Move pathfinding utils
